backend/requirements-wheels.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.5.0
python-multipart==0.0.6
PyJWT==2.8.0
python-dotenv==1.0.0
email-validator==2.1.0
gunicorn==21.2.0
typing-extensions==4.8.0